mlflow_airflow/kube/docker/requirements-test.txt
fastapi[standard]
pandas
logging
numpy 
pathlib 
scikit-learn
statsmodels
mlflow
uvicorn[standard]
pytest
requests
PyJWT
joblib
kubernetes
passlib